chore(trigger): move uprobe triggers to its own file
TODO: move all triggering logic to its own package
